Add support for downgrading a platform by forcing one of its leaves
This commits adds support for fixing the versions of all components of a virtual
platform by forcing one of its components. The engine automatically enforces all
members of the platform.
