+++ subprojects/dependency-management/src/integTest/groovy/org/gradle/integtests/resolve/alignment/AlignmentIntegrationTest.groovy
@@ -19,10 +19,8 @@ package org.gradle.integtests.resolve.alignment
 import org.gradle.integtests.fixtures.GradleMetadataResolveRunner
 import org.gradle.integtests.fixtures.RequiredFeature
 import org.gradle.integtests.fixtures.RequiredFeatures
-import org.gradle.integtests.fixtures.publish.RemoteRepositorySpec
-import org.gradle.integtests.resolve.AbstractModuleDependencyResolveTest
 
-class AlignmentIntegrationTest extends AbstractModuleDependencyResolveTest {
+class AlignmentIntegrationTest extends AbstractAlignmentSpec {
 
     def "should align leaves to the same version"() {
         repository {
@@ -821,167 +819,6 @@ class AlignmentIntegrationTest extends AbstractModuleDependencyResolveTest {
         }
     }
 
-    private void "a rule which infers module set from group and version"() {
-        buildFile << """
-            dependencies {
-                components.all(InferModuleSetFromGroupAndVersion)
-            }
-            
-            class InferModuleSetFromGroupAndVersion implements ComponentMetadataRule {
-                void execute(ComponentMetadataContext ctx) {
-                    ctx.details.with {
-                        belongsTo("\${id.group}:platform:\${id.version}")
-                    }
-                }
-            }
-        """
-    }
-
-    private void "align the 'org' group only"() {
-        buildFile << """
-            dependencies {
-                components.all(AlignOrgGroup)
-            }
-            
-            class AlignOrgGroup implements ComponentMetadataRule {
-                void execute(ComponentMetadataContext ctx) {
-                    ctx.details.with {
-                        if ('org' == id.group) {
-                           belongsTo("\${id.group}:platform:\${id.version}")
-                        }
-                    }
-                }
-            }
-        """
-    }
-
-    private void "align the 'org' group to 2 different virtual platforms"() {
-        buildFile << """
-            dependencies {
-                components.all(AlignOrgGroupTo2Platforms)
-            }
-            
-            class AlignOrgGroupTo2Platforms implements ComponentMetadataRule {
-                void execute(ComponentMetadataContext ctx) {
-                    ctx.details.with {
-                        if ('org' == id.group) {
-                           belongsTo("\${id.group}:platform:\${id.version}")
-                           belongsTo("\${id.group}:platform2:\${id.version}")
-                        }
-                    }
-                }
-            }
-        """
-    }
-
-    private 'a rule which declares that Groovy belongs to the Groovy and the Spring platforms'() {
-        buildFile << """
-            dependencies {
-                components.all(GroovyRule)
-            }
-            
-            class GroovyRule implements ComponentMetadataRule {
-                void execute(ComponentMetadataContext ctx) {
-                    ctx.details.with {
-                        if ('org.apache.groovy' == id.group) {
-                           belongsTo("org.apache.groovy:platform:\${id.version}")
-                           belongsTo("org.springframework:spring-platform:1.0")
-                        }
-                    }
-                }
-            }
-        """
-    }
-
-    final static Closure<Void> VIRTUAL_PLATFORM = {
-        expectGetMetadataMissing()
-        if (!GradleMetadataResolveRunner.experimentalResolveBehaviorEnabled) {
-            expectHeadArtifactMissing()
-        }
-    }
-
-    void expectAlignment(@DelegatesTo(value = AlignmentSpec, strategy = Closure.DELEGATE_FIRST) Closure<?> spec) {
-        def align = new AlignmentSpec()
-        spec.delegate = align
-        spec.resolveStrategy = Closure.DELEGATE_FIRST
-        spec()
-        repositoryInteractions {
-            align.applyTo(repoSpec)
-        }
-    }
-
-    static class AlignmentSpec {
-        final List<ModuleAlignmentSpec> specs = []
-        final Set<String> skipsPlatformMetadata = []
-
-        ModuleAlignmentSpec module(String name, @DelegatesTo(value=ModuleAlignmentSpec, strategy = Closure.DELEGATE_FIRST) Closure<?> config = null) {
-            def spec = new ModuleAlignmentSpec(name: name)
-            if (config) {
-                config.delegate = spec
-                config.resolveStrategy = Closure.DELEGATE_FIRST
-                config()
-            }
-            specs << spec
-            spec
-        }
-
-        void doesNotGetPlatform(String group = 'org', String name = 'platform', String version = '1.0') {
-            skipsPlatformMetadata << "$group:$name:$version"
-        }
-
-        void applyTo(RemoteRepositorySpec spec) {
-            Set<String> virtualPlatforms = [] as Set
-            Set<String> publishedPlatforms = [] as Set
-            Set<String> resolvesToVirtual = [] as Set
-
-            specs.each {
-                it.applyTo(spec)
-                if (it.virtualPlatforms) {
-                    it.seenVersions.each { v ->
-                        it.virtualPlatforms.each { vp ->
-                            virtualPlatforms << "${vp}:$v"
-                        }
-                    }
-                    it.virtualPlatforms.each { vp ->
-                        resolvesToVirtual << "${vp}:$it.alignsTo"
-                    }
-                }
-                if (it.publishedPlatforms) {
-                    def exactPlatforms = it.publishedPlatforms.findAll { it.count(':') == 2 }
-                    def inferredPlatforms = it.publishedPlatforms - exactPlatforms
-                    // for published platforms, we know there's no artifacts, so it's actually easier
-                    it.seenVersions.each { v ->
-                        inferredPlatforms.each { pp ->
-                            publishedPlatforms << "${pp}:$v"
-                        }
-                    }
-                    inferredPlatforms.each { pp ->
-                        publishedPlatforms << "${pp}:$it.alignsTo"
-                    }
-                    exactPlatforms.each { pp ->
-                        publishedPlatforms << pp
-                    }
-                }
-            }
-            virtualPlatforms.remove(resolvesToVirtual)
-            virtualPlatforms.removeAll(skipsPlatformMetadata)
-            resolvesToVirtual.removeAll(skipsPlatformMetadata)
-            publishedPlatforms.removeAll(skipsPlatformMetadata)
-            virtualPlatforms.each { p ->
-                spec."$p"(VIRTUAL_PLATFORM)
-            }
-            publishedPlatforms.each { p ->
-                spec."$p" {
-                    expectGetMetadata()
-                }
-            }
-            resolvesToVirtual.each {
-                spec."$it"(VIRTUAL_PLATFORM)
-            }
-        }
-
-    }
-
     @RequiredFeatures([
         // We only need to test one flavor
         @RequiredFeature(feature = GradleMetadataResolveRunner.GRADLE_METADATA, value = "true"),
@@ -1174,83 +1011,4 @@ class AlignmentIntegrationTest extends AbstractModuleDependencyResolveTest {
 
     }
 
-    static class ModuleAlignmentSpec {
-        String group = 'org'
-        String name
-        List<String> seenVersions = []
-        List<String> misses = []
-        String alignsTo
-        List<String> virtualPlatforms = []
-        List<String> publishedPlatforms = []
-
-        ModuleAlignmentSpec group(String group) {
-            this.group = group
-            this
-        }
-
-        ModuleAlignmentSpec name(String name) {
-            this.name = name
-            this
-        }
-
-        ModuleAlignmentSpec tries(String... versions) {
-            Collections.addAll(seenVersions, versions)
-            this
-        }
-
-        ModuleAlignmentSpec misses(String... versions) {
-            Collections.addAll(misses, versions)
-            this
-        }
-
-        ModuleAlignmentSpec alignsTo(String version) {
-            this.alignsTo = version
-            this
-        }
-
-        ModuleAlignmentSpec byVirtualPlatform(String group = 'org', String name = 'platform') {
-            virtualPlatforms << "${group}:${name}"
-            this
-        }
-
-        ModuleAlignmentSpec byPublishedPlatform(String group = 'org', String name = 'platform', String version = null) {
-            if (version) {
-                publishedPlatforms << "${group}:${name}:${version}"
-            } else {
-                publishedPlatforms << "${group}:${name}"
-            }
-            this
-        }
-
-
-        void applyTo(RemoteRepositorySpec spec) {
-            def moduleName = name
-            def alignedTo = alignsTo
-            def otherVersions = seenVersions
-            otherVersions.remove(alignedTo)
-            def missedVersions = misses
-            spec.group(group) {
-                module(moduleName) {
-                    if (alignedTo) {
-                        version(alignedTo) {
-                            expectResolve()
-                        }
-                    }
-                    otherVersions.each {
-                        version(it) {
-                            expectGetMetadata()
-                        }
-                    }
-                    missedVersions.each {
-                        version(it) {
-                            expectGetMetadataMissing()
-                            if (!GradleMetadataResolveRunner.experimentalResolveBehaviorEnabled) {
-                                expectHeadArtifactMissing()
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
 }
